Move general fuzzing code
